requirements.txt
Django==1.11.7
django-filter==1.1.0
django-jchart==0.4.2
djangorestframework==3.7.3
Fabric==1.14.0
psycopg2==2.7.3.2
